+++ packages/analytics/src/components/Map/interfaces/index.ts
@@ -60,6 +60,7 @@ export interface MapProps {
 	legends?: MapLegendConfig;
 	thematicLayers?: ThematicLayerConfig[];
 	earthEngineLayers?: EarthEngineLayerConfig[];
+	showPeriodTitle?: boolean;
 	periodSelection?: {
 		periods?: string[];
 		range?: { start: Date; end: Date };